setup separate ctl_lib
