+++ docker/entrypoint.sh
@@ -1,0 +1,4 @@
+#! /bin/sh
+
+rm -rf ./*
+cp /app/source/* ./
